% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximation.R
\name{approx}
\alias{approx}
\title{approx}
\usage{
approx(arcSources, arcTargets, numNodes, algorithm)
}
\arguments{
\item{arcSources, }{a vector corresponding to the source nodes of a graph's edges}

\item{arcTargets, }{a vector corresponding to the destination nodes of a graph's edges}

\item{numNodes, }{the number of nodes in the graph}

\item{algorithm}{Which algorithm to run. Choices include
"GrossoLocatelliPullanMc".
 GrossoLocatelliPullanMc is the default; see
<http://lemon.cs.elte.hu/pub/doc/1.3.1/a00619.html> for details
on the differences.}
}
\value{
a list containing two entries with 1) the clique size, and 2) the members of the clique.
}
\description{
Runs approximation algorthims for the maximal clique problem to find the largest complete subgraph (clique) in an undirected graph.
}
