% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Approximation-Algorithms}
\alias{Approximation-Algorithms}
\alias{GrossoLocatelliPullanMcRunner}
\title{Approximation Algorithms}
\usage{
GrossoLocatelliPullanMcRunner(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources, }{a vector corresponding to the source nodes of a graph's edges}

\item{arcTargets, }{a vector corresponding to the destination nodes of a graph's edges}

\item{numNodes, }{the number of nodes in the graph}
}
\value{
a List with 1) the clique size, and 2) the members of the clique
}
\description{
`GrossoLocatelliPullanMcRunner` runs the Grosso, Locatelli, and Pullan for solving the maximum clique problem
}
