% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_graph.R
\name{check_arc_map}
\alias{check_arc_map}
\title{Check arc maps for common errors. Doesn't check if the underlying graph is
correct - see `check_graph_vertices` for that purpose. Specifically checks
if the number of arcs in the graph is the same as the number in the map.}
\usage{
check_arc_map(arcSources, arcTargets, mapTest, numNodes)
}
\arguments{
\item{arcSources}{Arc sources}

\item{arcTargets}{Arc Targets}

\item{mapTest}{Arc Map to Test}

\item{numNodes}{Number of nodes}
}
\value{
TRUE if no errors are found
export
}
\description{
Check arc maps for common errors. Doesn't check if the underlying graph is
correct - see `check_graph_vertices` for that purpose. Specifically checks
if the number of arcs in the graph is the same as the number in the map.
}
