% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_graph.R
\name{check_graph_arcs}
\alias{check_graph_arcs}
\title{Check arcs for common errors.}
\usage{
check_graph_arcs(arcSources, arcTargets)
}
\arguments{
\item{arcSources}{Arc sources}

\item{arcTargets}{Arc Targets}
}
\value{
TRUE if no errors are found
export
}
\description{
Check arcs for common errors.
}
