% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_graph.R
\name{check_node}
\alias{check_node}
\title{Check node definitions for common errors}
\usage{
check_node(testNode, numNodes)
}
\arguments{
\item{testNode}{Node to test}

\item{numNodes}{Number of nodes}
}
\value{
TRUE if no errors are found
export
}
\description{
Check node definitions for common errors
}
