% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_graph.R
\name{check_node_map}
\alias{check_node_map}
\title{Check node maps for common errors. Doesn't check if the underlying graph is
correct - see `check_graph_vertices` for that purpose. Specifically checks
if the number of nodes is inconsistent.}
\usage{
check_node_map(mapTest, numNodes)
}
\arguments{
\item{mapTest}{Node Map to Test}

\item{numNodes}{Number of nodes}
}
\value{
TRUE if no errors are found
export
}
\description{
Check node maps for common errors. Doesn't check if the underlying graph is
correct - see `check_graph_vertices` for that purpose. Specifically checks
if the number of nodes is inconsistent.
}
