% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ChristofidesRunner}
\alias{ChristofidesRunner}
\title{ChristofidesRunner}
\usage{
ChristofidesRunner(
  arcSources,
  arcTargets,
  arcDistances,
  numNodes,
  defaultEdgeWeight = 999999L
)
}
\description{
ChristofidesRunner
}
