% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Connectivity-Algorithms-2}
\alias{Connectivity-Algorithms-2}
\alias{CountBiEdgeConnectedComponentsRunner}
\alias{CountConnectedComponentsRunner}
\alias{CountBiNodeConnectedComponentsRunner}
\alias{CountStronglyConnectedComponentsRunner}
\title{Connectivity Algorithms - Integer Return}
\usage{
CountBiEdgeConnectedComponentsRunner(arcSources, arcTargets, numNodes)

CountConnectedComponentsRunner(arcSources, arcTargets, numNodes)

CountBiNodeConnectedComponentsRunner(arcSources, arcTargets, numNodes)

CountStronglyConnectedComponentsRunner(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources, }{a vector corresponding to the source nodes of a graph's edges}

\item{arcTargets, }{a vector corresponding to the destination nodes of a graph's edges}

\item{numNodes, }{the number of nodes in the graph}
}
\value{
an integer, counting some property of a graph
}
\description{
`CountBiEdgeConnectedComponents` returns how many Bi-Edge Connected Components a graph has.

`CountConnectedComponents` returns how many Connected Components a graph has.

`CountBiNodeConnectedComponents` returns how many Bi-Node Connected Components a graph has.

`CountStronglyConnectedComponents` returns how many Strongly Connected Components a graph has.
}
