% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{CountBiEdgeConnectedComponents}
\alias{CountBiEdgeConnectedComponents}
\title{CountBiEdgeConnectedComponents}
\usage{
CountBiEdgeConnectedComponents(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's
edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's
edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
An integer defining the number of bi-node connected components
}
\description{
Counts the number of bi-edge connected components in an undirected graph.
}
