% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CountBiNodeConnectedComponents}
\alias{CountBiNodeConnectedComponents}
\title{CountBiNodeConnectedComponents}
\usage{
CountBiNodeConnectedComponents(arcSources, arcTargets, numNodes)
}
\description{
CountBiNodeConnectedComponents
}
