% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CountStronglyConnectedComponents}
\alias{CountStronglyConnectedComponents}
\title{CountStronglyConnectedComponents}
\usage{
CountStronglyConnectedComponents(arcSources, arcTargets, numNodes)
}
\description{
CountStronglyConnectedComponents
}
