% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FindBiEdgeConnectedCutEdges}
\alias{FindBiEdgeConnectedCutEdges}
\title{FindBiEdgeConnectedCutEdges}
\usage{
FindBiEdgeConnectedCutEdges(arcSources, arcTargets, numNodes)
}
\description{
FindBiEdgeConnectedCutEdges
}
