% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FindBiNodeConnectedComponents}
\alias{FindBiNodeConnectedComponents}
\title{FindBiNodeConnectedComponents}
\usage{
FindBiNodeConnectedComponents(arcSources, arcTargets, numNodes)
}
\description{
FindBiNodeConnectedComponents
}
