% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{FindBiNodeConnectedCutNodes}
\alias{FindBiNodeConnectedCutNodes}
\title{FindBiNodeConnectedCutNodes}
\usage{
FindBiNodeConnectedCutNodes(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's
edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's
edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
A list containing the cut nodes.
}
\description{
Finds the bi-node-connected cut nodes in an undirected graph.
}
