% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FindBiNodeConnectedNodes}
\alias{FindBiNodeConnectedNodes}
\title{FindBiNodeConnectedNodes}
\usage{
FindBiNodeConnectedNodes(arcSources, arcTargets, numNodes)
}
\description{
FindBiNodeConnectedNodes
}
