% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FindStronglyConnectedCutArcs}
\alias{FindStronglyConnectedCutArcs}
\title{FindStronglyConnectedCutArcs}
\usage{
FindStronglyConnectedCutArcs(arcSources, arcTargets, numNodes)
}
\description{
FindStronglyConnectedCutArcs
}
