% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getAndcheckTopologicalSort}
\alias{getAndcheckTopologicalSort}
\title{Gets the topological sort on a list of graphs}
\usage{
getAndcheckTopologicalSort(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{A vector of integers containing the source vertices of the edges}

\item{arcTargets}{A vector of integers containing the destination verctices of the edges}

\item{numNodes}{The number of nodes in the graph.}
}
\value{
An R List containing (1) A boolean stating if the graph is a dag, and (2) an std::vector of length numNodes, containing the index of vertex i in the ordering at location i
}
\description{
Gets the topological sort on a list of graphs
}
