% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getBipartitePartitions}
\alias{getBipartitePartitions}
\title{Acquires bipartite partitions}
\usage{
getBipartitePartitions(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{A vector of integers containing the source vertices of the edges}

\item{arcTargets}{A vector of integers containing the destination verctices of the edges}

\item{numNodes}{The number of nodes in the graph.}
}
\description{
Acquires bipartite partitions
}
