% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getTopologicalSort}
\alias{getTopologicalSort}
\title{Gets the topological sort on a list of graphs}
\usage{
getTopologicalSort(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{A vector of integers containing the source vertices of the edges}

\item{arcTargets}{A vector of integers containing the destination verctices of the edges}

\item{numNodes}{The number of nodes in the graph.}
}
\value{
An R Numeric Vector of length numNodes, containing the index of vertex i in the ordering at location i.
}
\description{
Gets the topological sort on a list of graphs
}
