% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Graph-Converter}
\alias{Graph-Converter}
\alias{GraphCompatabilityConverter}
\title{Compatible Graph Converter}
\usage{
GraphCompatabilityConverter(nodesList, arcSources, arcTargets)
}
\arguments{
\item{nodesList, }{a vector of all the nodes in the graph}

\item{arcSources, }{a vector corresponding to the source nodes of a graph's edges}

\item{arcTargets, }{a vector corresponding to the destination nodes of a graph's edges}
}
\value{
A list containing four entries: 1) rlemon compatible nodes, 2) arcSources which match the converted nodes 3) arcTargets which match the converted nodes, 4) a mapping of the original and converted nodes
}
\description{
'GraphCompatabilityConverter' returns a compatible version of the graph which can be used with the rest of rlemon's interfaces. rlemon's interface requires all node inputs to be 0-indexed.
}
