% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Graph-Search-Algorithms-1}
\alias{Graph-Search-Algorithms-1}
\alias{BfsRunner}
\alias{DfsRunner}
\title{Graph Search Algorithms}
\usage{
BfsRunner(arcSources, arcTargets, numNodes, startNode = -1L, endNode = -1L)

DfsRunner(arcSources, arcTargets, numNodes, startNode = -1L, endNode = -1L)
}
\arguments{
\item{arcSources, }{a vector corresponding to the source nodes of a graph's edges}

\item{arcTargets, }{a vector corresponding to the destination nodes of a graph's edges}

\item{numNodes, }{the number of nodes in the graph}

\item{startNode, }{the start node of the path}

\item{endNode, }{the end node of the path}
}
\value{
A list containing three entries: 1) the predecessor of each vertex in its shortest path, 2) the distances from each node to the startNode , 3) a list containing if a node was reached or not
}
\description{
`BfsRunner` calculates the shortest path using Breath-First-Search

`DfsRunner` calculates the shortest path using Depth-First-Search
}
