% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Graph-Search-Algorithms-2}
\alias{Graph-Search-Algorithms-2}
\alias{MaxCardinalitySearchRunner}
\title{Graph Search Algorithms}
\usage{
MaxCardinalitySearchRunner(
  arcSources,
  arcTargets,
  arcCapacities,
  numNodes,
  startNode = -1L
)
}
\arguments{
\item{arcSources, }{a vector corresponding to the source nodes of a graph's edges}

\item{arcTargets, }{a vector corresponding to the destination nodes of a graph's edges}

\item{arcCapacities, }{a vector corresponding to the distances of a graph's edges}

\item{numNodes, }{the number of nodes in the graph}

\item{startNode, }{the start node of the path}
}
\value{
A list containing two entries: 1) the cardinality of each node , 2) a list containing if a node was reached or not
}
\description{
`MaxCardinalitySearchRunner` runs the Maximum Cardinality Search Algorithm
}
