% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GreedyTSPRunner}
\alias{GreedyTSPRunner}
\title{GreedyTSPRunner}
\usage{
GreedyTSPRunner(
  arcSources,
  arcTargets,
  arcDistances,
  numNodes,
  defaultEdgeWeight = 999999L
)
}
\description{
GreedyTSPRunner
}
