% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{HartmannOrlinMmcRunner}
\alias{HartmannOrlinMmcRunner}
\title{HartmannOrlinMmcRunner}
\usage{
HartmannOrlinMmcRunner(arcSources, arcTargets, arcDistances, numNodes)
}
\description{
HartmannOrlinMmcRunner
}
