% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{InsertionTSPRunner}
\alias{InsertionTSPRunner}
\title{InsertionTSPRunner}
\usage{
InsertionTSPRunner(
  arcSources,
  arcTargets,
  arcDistances,
  numNodes,
  defaultEdgeWeight = 999999L
)
}
\description{
InsertionTSPRunner
}
