% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{IsAcyclic}
\alias{IsAcyclic}
\title{IsAcylic}
\usage{
IsAcyclic(arcSources, arcTargets, numNodes, problem = "isAcyclic")
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's edges}

\item{numNodes}{The number of nodes in the graph}

\item{problem}{What type of problem you want to run. Choices include
"isAcylic" 
where "isAcylic" is the default.
See <http://lemon.cs.elte.hu/pub/doc/1.3.1/a00616.html> 
for details on the differences.}
}
\value{
A boolean stating if the graph is acyclic
}
\description{
Checks if an undirected graph is acyclic
}
