% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{IsBiEdgeConnected}
\alias{IsBiEdgeConnected}
\title{IsBiEdgeConnected}
\usage{
IsBiEdgeConnected(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's
edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's
edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
A boolean stating if the graph is bi-edge connected
}
\description{
Checks if an undirected graph is bi-edge connected
}
