% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{IsEulerian}
\alias{IsEulerian}
\title{IsEulerian}
\usage{
IsEulerian(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
A boolean stating if the graph is eulerian
}
\description{
Checks if a graph is eulerian. Works for both undirected and directed graphs.
}
