% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Maximum Flow Algorithms 2}
\alias{Maximum Flow Algorithms 2}
\alias{CirculationRunner}
\title{Maximum  Flow Algorithms - Network Circulation}
\usage{
CirculationRunner(
  arcSources,
  arcTargets,
  arcLowerBound,
  arcUpperBound,
  nodeSupplies,
  numNodes
)
}
\arguments{
\item{arcSources, }{a vector corresponding to the source nodes of a graph's edges}

\item{arcTargets, }{a vector corresponding to the destination nodes of a graph's edges}

\item{arcLowerBound, }{a vector corresponding to the lower-bound capacities of nodes of a graph's edges}

\item{arcUpperBound, }{a vector corresponding to the upper-bound capacities of nodes of a graph's edges}

\item{nodeSupplies, }{a vector corresponding to the supplies of each node of the graph.}

\item{numNodes, }{the number of nodes in the graph}
}
\value{
A list containing three entries: 1) A list corresponding to the flows of arcs in the graph, 2) A list of the graph's barrier nodes, and 3) the total cost of the flows in the graph, i.e. the maxflow value
}
\description{
`PreflowRunner` runs the Preflow Algorithm to calculate the maximum flow.
}
