% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Maximum Flow Algorithms}
\alias{Maximum Flow Algorithms}
\alias{PreflowRunner}
\alias{EdmondsKarpRunner}
\title{Maximum  Flow Algorithms}
\usage{
PreflowRunner(
  arcSources,
  arcTargets,
  arcDistances,
  sourceNode,
  destinationNode,
  numNodes
)

EdmondsKarpRunner(
  arcSources,
  arcTargets,
  arcDistances,
  sourceNode,
  destinationNode,
  numNodes
)
}
\arguments{
\item{arcSources, }{a vector corresponding to the source nodes of a graph's edges}

\item{arcTargets, }{a vector corresponding to the destination nodes of a graph's edges}

\item{arcDistances, }{a vector corresponding to the capacities of nodes of a graph's edges}

\item{sourceNode, }{the source node}

\item{destinationNode, }{the destination node}

\item{numNodes, }{the number of nodes in the graph}
}
\value{
A list containing three entries: 1) A list corresponding to the flows of arcs in the graph, and 2) A list of cut-values of the graph's nodes.
}
\description{
`PreflowRunner` runs the Preflow Algorithm to calculate the maximum flow.

`EdmondsKarpRunner` runs the EdmondsKarp Algorithm to calculate the maximum flow.
}
