% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MaximumCardinalityMatchingRunner}
\alias{MaximumCardinalityMatchingRunner}
\title{MaximumCardinalityMatchingRunner}
\usage{
MaximumCardinalityMatchingRunner(arcSources, arcTargets, numNodes)
}
\description{
MaximumCardinalityMatchingRunner
}
