% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MaximumWeightMatchingRunner}
\alias{MaximumWeightMatchingRunner}
\title{MaximumWeightMatchingRunner}
\usage{
MaximumWeightMatchingRunner(arcSources, arcTargets, arcWeights, numNodes)
}
\description{
MaximumWeightMatchingRunner
}
