% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MaximumWeightPerfectMatchingRunner}
\alias{MaximumWeightPerfectMatchingRunner}
\title{MaximumWeightPerfectMatchingRunner}
\usage{
MaximumWeightPerfectMatchingRunner(
  arcSources,
  arcTargets,
  arcWeights,
  numNodes
)
}
\description{
MaximumWeightPerfectMatchingRunner
}
