% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxmatching.R
\name{maxmatching}
\alias{maxmatching}
\title{MinCostFlow}
\usage{
maxmatching(
  arcSources,
  arcTargets,
  numNodes,
  problem = "MaxWeightedMatching",
  arcWeights = NULL
)
}
\arguments{
\item{arcSources, }{a vector corresponding to the source nodes of a graph's edges}

\item{arcTargets, }{a vector corresponding to the destination nodes of a graph's edges}

\item{numNodes, }{the number of nodes in the graph}

\item{problem}{What type of problem you want to run. Choices include
"MaxMatching", "MaxWeightedMatching", "MaxWeightedPerfectMatching", 
"MaxFractionalMatching", "MaxWeightedFractionalMatching", "MaxWeightedPerfectFractionalMatching"
where "MaxWeightedMatching" is the default.
See <http://lemon.cs.elte.hu/pub/doc/1.3.1/a00615.html> 
for details on the differences.}

\item{arcWeights, }{a vector corresponding to the weights of a graph's edges}
}
\value{
A list containing two entries: 1) The matching value, 2) The edges of the final graph, in a List of List of (node, node) pairs
}
\description{
Runs maxflow algorithms on graphs.
}
