% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mincostflow.R
\name{mincostflow}
\alias{mincostflow}
\title{MinCostFlow}
\usage{
mincostflow(
  arcSources,
  arcTargets,
  arcCapacities,
  arcCosts,
  nodeSupplies,
  numNodes,
  algorithm
)
}
\arguments{
\item{arcSources}{Arc sources}

\item{arcTargets}{Arc targets}

\item{arcCapacities}{Arc capacities}

\item{arcCosts}{Arc costs}

\item{nodeSupplies}{Node supplies}

\item{numNodes}{Number of nodes}

\item{algorithm}{Which algorithm to run. Choices include
"NetworkSimplex", "CostScaling", "CapacityScaling",
"CycleCancelling". NetworkSimplex is the default; see
<http://lemon.cs.elte.hu/pub/doc/1.3.1/a00612.html> for details
on the differences.}
}
\value{

}
\description{
Runs mincostflow algorithms on graphs.
}
