% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Minimum Cost Aborescence Algorithms}
\alias{Minimum Cost Aborescence Algorithms}
\alias{MinCostArborescenceRunner}
\title{Aborescense Algorithms}
\usage{
MinCostArborescenceRunner(
  arcSources,
  arcTargets,
  arcDistances,
  sourceNode,
  numNodes
)
}
\arguments{
\item{arcSources, }{a vector corresponding to the source nodes of a graph's edges}

\item{arcTargets, }{a vector corresponding to the destination nodes of a graph's edges}

\item{arcDistances, }{a vector corresponding to the distances of nodes of a graph's edges}

\item{sourceNode, }{the source node}

\item{numNodes, }{the number of nodes in the graph}
}
\value{
A list containing three entries: 1) Two vectors corresponding the source and target nodes of the edges in the trre, and 2) the total minimum spanning tree value.
}
\description{
Aborescense Algorithms
}
