% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Minimum-Spanning-Tree-Algorithms}
\alias{Minimum-Spanning-Tree-Algorithms}
\alias{KruskalRunner}
\title{MST Algorithms}
\usage{
KruskalRunner(arcSources, arcTargets, arcDistances, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's edges}

\item{arcDistances}{Vector corresponding to the distances of nodes of a graph's edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
A list containing three entries: 1) Two vectors corresponding the source and target nodes of the edges in the tree, and 2) the total minimum spanning tree value.
}
\description{
MST Algorithms
}
