% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimummeancycle.R
\name{minimummeancycle}
\alias{minimummeancycle}
\title{MinimumMeanCycle}
\usage{
minimummeancycle(arcSources, arcTargets, arcDistances, numNodes, algorithm)
}
\arguments{
\item{arcSources, }{a vector corresponding to the source nodes of a graph's edges}

\item{arcTargets, }{a vector corresponding to the destination nodes of a graph's edges}

\item{arcDistances, }{a vector corresponding to the distances of a graph's edges}

\item{numNodes, }{the number of nodes in the graph}

\item{algorithm}{Which algorithm to run. Choices include
"Howard", "Karp", "HartmannOrlin"
where "Howard" is the default.
See <http://lemon.cs.elte.hu/pub/doc/1.3.1/a00614.html> 
for details on the differences.}
}
\value{
A list containing two entries: 1) A vector containing the costs of each edge in the MMC cycle, and 2) the nodes in the MMC cycle.
}
\description{
Runs MMC (Minimum Mean Cycle) algorithms on graphs.
}
