% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimumspanningtree.R
\name{minimummspanningtree}
\alias{minimummspanningtree}
\title{MinimumSpanningTree}
\usage{
minimummspanningtree(
  arcSources,
  arcTargets,
  arcDistances,
  numNodes,
  algorithm = "Kruskal",
  sourceNode = NULL
)
}
\arguments{
\item{arcSources, }{a vector corresponding to the source nodes of a graph's edges}

\item{arcTargets, }{a vector corresponding to the destination nodes of a graph's edges}

\item{arcDistances, }{a vector corresponding to the distances of nodes of a graph's edges}

\item{numNodes, }{the number of nodes in the graph}

\item{algorithm}{Which algorithm to run. Choices include
"Kruskal", "MinCostArborescence"
where "Kruskal" is the default.
See <http://lemon.cs.elte.hu/pub/doc/1.3.1/a00610.html> 
for details on the differences.}

\item{sourceNode, }{the source node (only used for MinCostArborescence)}
}
\value{
A list containing three entries: 1) Two vectors corresponding the source and target nodes of the edges in the tree, and 2) the total minimum spanning tree value.
}
\description{
Runs MST (Minimum Spanning Tree) algorithms on graphs.
}
