% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NearestNeighborTSPRunner}
\alias{NearestNeighborTSPRunner}
\title{NearestNeighborTSPRunner}
\usage{
NearestNeighborTSPRunner(
  arcSources,
  arcTargets,
  arcDistances,
  numNodes,
  defaultEdgeWeight = 999999L
)
}
\description{
NearestNeighborTSPRunner
}
