% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Opt2TSPRunner}
\alias{Opt2TSPRunner}
\title{Opt2TSPRunner}
\usage{
Opt2TSPRunner(
  arcSources,
  arcTargets,
  arcDistances,
  numNodes,
  defaultEdgeWeight = 999999L
)
}
\description{
Opt2TSPRunner
}
