% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Planar-Embedding-Algorithms-2}
\alias{Planar-Embedding-Algorithms-2}
\alias{PlanarEmbeddingRunner}
\title{Planar Embedding Algorithms}
\usage{
PlanarEmbeddingRunner(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources, }{a vector corresponding to the source nodes of a graph's edges}

\item{arcTargets, }{a vector corresponding to the destination nodes of a graph's edges}

\item{numNodes, }{the number of nodes in the graph}
}
\value{
A list containing the following 1) A boolean if the graph is planar or not, 2) start/end node lists for the vertices in the order of the Planar Embedding (if planar), 3) start/end node lists for the arcs in the kuratowski subdivision (if not planar)
}
\description{
`PlanarEmbeddingRunner` runs the Planar Embedding Algorithm to prove that some graph is planar or not.
}
