% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planarityembedding.R
\name{PlanarChecking}
\alias{PlanarChecking}
\title{PlanarChecking}
\usage{
PlanarChecking(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
A boolean stating if the graph is planar or not.
}
\description{
Checks if an undirected graph is planar and returns a boolean if it is or not.
}
