% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planarityembedding.R
\name{PlanarDrawing}
\alias{PlanarDrawing}
\title{PlanarDrawing}
\usage{
PlanarDrawing(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's
edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's
edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
A list of 1) a boolean of if the graph is planar, 2) the
  x-coordinate of the planar embedding, 3) the y-coordinate of the planar
  embedding
}
\description{
Checks if a graph is planar and returns the coloring of the graph
See 'https://lemon.cs.elte.hu/pub/doc/1.3.1/a00307.html'
for more information.
}
