% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Shortest-Path-Algorithms-1}
\alias{Shortest-Path-Algorithms-1}
\alias{DijkstraRunner}
\alias{BellmanFordRunner}
\title{Shortest Path Algorithms - Distance}
\usage{
DijkstraRunner(arcSources, arcTargets, arcDistances, numNodes, startNode)

BellmanFordRunner(arcSources, arcTargets, arcDistances, numNodes, startNode)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's edges}

\item{arcDistances}{Vector corresponding to the distances of a graph's edges}

\item{numNodes}{The number of nodes in the graph}

\item{startNode}{The start node of the path}
}
\value{
A list containing two entries: 1) the distances from each node to the startNode and 2) the predecessor of each vertex in its shortest path.
}
\description{
`DijkstraRunner` calculates the shortest path from the start node to each node in the graph and returns a list containing 1) the distances from each node to the startNode and 2) the predecessor of each vertex in its shortest path.

`BellmanFordRunner` calculates the shortest path from the start node to each node in the graph and returns a list containing 1) the distances from each node to the startNode and 2) the predecessor of each vertex in its shortest path.
}
