% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Shortest-Path-Algorithms-2}
\alias{Shortest-Path-Algorithms-2}
\alias{SuurballeRunner}
\title{Shortest Path Algorithms - Paths}
\usage{
SuurballeRunner(
  arcSources,
  arcTargets,
  arcDistances,
  numNodes,
  startNode,
  endNode
)
}
\arguments{
\item{arcSources, }{a vector corresponding to the source nodes of a graph's edges}

\item{arcTargets, }{a vector corresponding to the destination nodes of a graph's edges}

\item{arcDistances, }{a vector corresponding to the distances of a graph's edges}

\item{numNodes, }{the number of nodes in the graph}

\item{startNode, }{the start node of the path}

\item{endNode, }{the end node of the path}
}
\value{
A list containing two entries: 1) the number of arc disjoint paths from the start node to the end node and 2) a list of paths found.
}
\description{
`SuurballeRunner` calculates the shortest path between a specificed start and end node and returns a list containing  1) the number of arc disjoint paths from the start node to the end node and 2) a list of paths found.
}
