% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortestpath.R
\name{shortestpath}
\alias{shortestpath}
\title{ShortestPath}
\usage{
shortestpath(
  arcSources,
  arcTargets,
  arcDistances,
  sourceNode,
  numNodes,
  algorithm = "Dijkstra",
  destNode = -1
)
}
\arguments{
\item{arcSources, }{a vector corresponding to the source nodes of a graph's edges}

\item{arcTargets, }{a vector corresponding to the destination nodes of a graph's edges}

\item{arcDistances, }{a vector corresponding to the distances of a graph's edges}

\item{sourceNode, }{the start node of the path}

\item{numNodes, }{the number of nodes in the graph}

\item{algorithm}{Which algorithm to run. Choices include
"Suurballe", "Dijkstra", "BellmanFord"
where "Dijkstra" is the default.
See <http://lemon.cs.elte.hu/pub/doc/1.3.1/a00608.html>
for details on the differences.}

\item{destNode, }{the end node of the path}
}
\value{
"BellmanFord" and "Dijktra" return a list containing two entries: 1) the distances from each node to the startNode and 2) the predecessor of each vertex in its shortest path. "Suurballe" returns A list containing two entries: 1) the number of paths from the start node to the end node and 2) a list of paths found.
}
\description{
Runs maxflow algorithms on graphs.
}
