% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_graph_example}
\alias{small_graph_example}
\title{A small network graph example}
\format{
A list of length 5.
}
\usage{
small_graph_example
}
\description{
A small network graph example
}
\keyword{dataset}
