% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_graph.R
\name{check_graph}
\alias{check_graph}
\title{Check graph inputs for common issues. Specifically checks if
arc source, target, capacity, and cost arrays have the same length.
Additionally, checks if all arc sources and targets are integral.}
\usage{
check_graph(
  arcSources,
  arcTargets,
  arcCapacities,
  arcCosts,
  nodeSupplies,
  numNodes
)
}
\arguments{
\item{arcSources}{Arc sources}

\item{arcTargets}{Arc targets}

\item{arcCapacities}{Arc capacities}

\item{arcCosts}{Arc costs}

\item{nodeSupplies}{Node supplies}

\item{numNodes}{Number of nodes}
}
\value{
TRUE if no errors are found.
}
\description{
Check graph inputs for common issues. Specifically checks if
arc source, target, capacity, and cost arrays have the same length.
Additionally, checks if all arc sources and targets are integral.
}
