% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{CountBiNodeConnectedComponents}
\alias{CountBiNodeConnectedComponents}
\title{CountBiNodeConnectedComponents}
\usage{
CountBiNodeConnectedComponents(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's
edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's
edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
An integer defining the number of bi-node connected components
}
\description{
Counts the number of bi-node connected components in an undirected graph.
See 'https://lemon.cs.elte.hu/pub/doc/1.3.1/a00616.html#gaf7c5744b2175210b8ea67897aaa27885'
for more information.
}
