% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CountConnectedComponents}
\alias{CountConnectedComponents}
\title{CountConnectedComponents}
\usage{
CountConnectedComponents(arcSources, arcTargets, numNodes)
}
\description{
CountConnectedComponents
}
