% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{CountConnectedComponents}
\alias{CountConnectedComponents}
\title{Count the Number of Connected Components}
\usage{
CountConnectedComponents(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's
edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's
edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
An integer defining the number of connected components
}
\description{
The connected components are the classes of an equivalence relation on the
nodes of an undirected graph. Two nodes are in the same class if they are
connected with a path.
}
\details{
See \url{https://lemon.cs.elte.hu/pub/doc/1.3.1/a00616.html#ga33a9d9d4803cb15e83568b2526e978a5}
for more information.
}
