% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{CountStronglyConnectedComponents}
\alias{CountStronglyConnectedComponents}
\title{CountStronglyConnectedComponents}
\usage{
CountStronglyConnectedComponents(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's
edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's
edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
An integer defining the number of strongly connected components
}
\description{
Counts the number of strongly connected components in a directed graph.
See 'https://lemon.cs.elte.hu/pub/doc/1.3.1/a00616.html#gad30bc47dfffb78234eeee903cb3766f4'
for more information.
}
