% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FindBiEdgeConnectedComponents}
\alias{FindBiEdgeConnectedComponents}
\title{FindBiEdgeConnectedComponents}
\usage{
FindBiEdgeConnectedComponents(arcSources, arcTargets, numNodes)
}
\description{
FindBiEdgeConnectedComponents
}
