% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{FindBiEdgeConnectedComponents}
\alias{FindBiEdgeConnectedComponents}
\title{FindBiEdgeConnectedComponents}
\usage{
FindBiEdgeConnectedComponents(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's
edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's
edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
A list containing the node id of each bi-edge-connected component.
}
\description{
Finds the bi-edge-connected components of an undirected graph.
See 'https://lemon.cs.elte.hu/pub/doc/1.3.1/a00616.html#ga76c1fdd1881d21677507100b7e96c983'
for more information.
}
