% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{FindBiEdgeConnectedCutEdges}
\alias{FindBiEdgeConnectedCutEdges}
\title{FindBiEdgeConnectedCutEdges}
\usage{
FindBiEdgeConnectedCutEdges(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's
edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's
edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
A list containing 1) A list of cut arc sources, and 2) A list of cut
  arc destinations.
}
\description{
Finds the bi-edge-connected cut edges in an undirected graph.
See 'https://lemon.cs.elte.hu/pub/doc/1.3.1/a00616.html#ga58d444eba448c5f1a53539bd1b69636e'
for more information.
}
