% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{FindBiNodeConnectedComponents}
\alias{FindBiNodeConnectedComponents}
\title{FindBiNodeConnectedComponents}
\usage{
FindBiNodeConnectedComponents(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's
edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's
edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
A list containing the arc id of each cut arc.
}
\description{
Finds the bi-node-connected components of an undirected graph.
See 'https://lemon.cs.elte.hu/pub/doc/1.3.1/a00616.html#ga9d70526ab54e10b4b6fe3762af8675dd'
for more information.
}
