% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{FindBiNodeConnectedCutNodes}
\alias{FindBiNodeConnectedCutNodes}
\title{Find Bi-Node-Connected Cut Nodes}
\usage{
FindBiNodeConnectedCutNodes(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's
edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's
edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
A vector containing the cut nodes.
}
\description{
The bi-node-connected components are the classes of an equivalence relation
on the edges of a undirected graph. Two edges are in the same class if they
are on same circle.
}
\details{
See \url{https://lemon.cs.elte.hu/pub/doc/1.3.1/a00616.html#ga31461f33a748327ea3ef2a3199ffb6c7}
for more information.
}
