% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{FindConnectedComponents}
\alias{FindConnectedComponents}
\title{Find Connected Components}
\usage{
FindConnectedComponents(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's
edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's
edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
A vector containing the node id of each connected component.
}
\description{
The connected components are the classes of an equivalence relation on the
nodes of an undirected graph. Two nodes are in the same class if they are
connected with a path.
}
\details{
See \url{https://lemon.cs.elte.hu/pub/doc/1.3.1/a00616.html#gaa467a3e0a8c2e9e762650fd01fadff89}
for more information.
}
