% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{FindConnectedComponents}
\alias{FindConnectedComponents}
\title{FindConnectedComponents}
\usage{
FindConnectedComponents(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's
edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's
edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
A list containing the node id of each connected component.
}
\description{
Finds the connected components in an undirected graph.
See 'https://lemon.cs.elte.hu/pub/doc/1.3.1/a00616.html#gaa467a3e0a8c2e9e762650fd01fadff89'
for more information.
}
