% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{FindStronglyConnectedComponents}
\alias{FindStronglyConnectedComponents}
\title{FindStronglyConnectedComponents}
\usage{
FindStronglyConnectedComponents(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's
edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's
edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
A list containing the node id of each strongly connected component.
}
\description{
Finds the strongly connected components in a directed graph.
}
