% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{FindStronglyConnectedComponents}
\alias{FindStronglyConnectedComponents}
\title{Find Strongly Connected Components}
\usage{
FindStronglyConnectedComponents(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's
edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's
edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
A vector containing the node id of each strongly connected component.
}
\description{
The strongly connected components are the classes of an equivalence relation
on the nodes of a directed graph. Two nodes are in the same class if they
are connected with directed paths in both direction.
}
\details{
See \url{https://lemon.cs.elte.hu/pub/doc/1.3.1/a00616.html#ga46f8c22f3e2989c4689faa4c46ec9436}
for more information.
}
