% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{FindStronglyConnectedCutArcs}
\alias{FindStronglyConnectedCutArcs}
\title{FindStronglyConnectedCutArcs}
\usage{
FindStronglyConnectedCutArcs(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's
edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's
edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
A list containing 1) A list of cut arc sources, and 2) A list of cut
  arc destinations.
}
\description{
Finds the cut arcs of strongly connected components in a directed graph.
See 'https://lemon.cs.elte.hu/pub/doc/1.3.1/a00616.html#gad7af5c3a97453e37f251f0e86dbb83db'
for more information.
}
