% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{getAndCheckTopologicalSort}
\alias{getAndCheckTopologicalSort}
\title{GetAndCheckTopologicalSort}
\usage{
getAndCheckTopologicalSort(
  arcSources,
  arcTargets,
  numNodes,
  problem = "getAndCheckTopologicalSort"
)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's edges}

\item{numNodes}{The number of nodes in the graph}

\item{problem}{What type of problem you want to run. Choices include
"getAndCheckTopologicalSort" 
where "getAndCheckTopologicalSort" is the default.
See <http://lemon.cs.elte.hu/pub/doc/1.3.1/a00616.html> 
for details on the differences.}
}
\value{
A list containing 1) A boolean stating if the graph is a dag, and 2) A vector of length numNodes, containing the index of vertex i in the ordering at location i
}
\description{
Checks if a directed graph is a DAG and returns the topological order
}
