% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{GetAndCheckTopologicalSort}
\alias{GetAndCheckTopologicalSort}
\title{GetAndCheckTopologicalSort}
\usage{
GetAndCheckTopologicalSort(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's
edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's
edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
A list containing 1) A boolean stating if the graph is a dag, and 2)
  A vector of length numNodes, containing the index of vertex i in the
  ordering at location i
}
\description{
Checks if a directed graph is a DAG and returns the topological order.
See 'https://lemon.cs.elte.hu/pub/doc/1.3.1/a00616.html#gafc2cb20cf3859f157c0e12da7f310bb3'
for more information.
}
