% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{getBipartitePartitions}
\alias{getBipartitePartitions}
\title{GetBipartitePartitions}
\usage{
getBipartitePartitions(
  arcSources,
  arcTargets,
  numNodes,
  problem = "getBipartitePartitions"
)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's edges}

\item{numNodes}{The number of nodes in the graph}

\item{problem}{What type of problem you want to run. Choices include
"getBipartitePartitions" 
where "getBipartitePartitions" is the default.
See <http://lemon.cs.elte.hu/pub/doc/1.3.1/a00616.html> 
for details on the differences.}
}
\value{
A list containing 1) A boolean stating if the graph is bipartite, and 2) A list of length numNodes, containing the partition for each node
}
\description{
Checks if an undirected graph is bipartite and returns the bipartite partitions
}
