% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{GetTopologicalSort}
\alias{GetTopologicalSort}
\title{Sorts Nodes into Topological Order}
\usage{
GetTopologicalSort(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's
edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's
edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
A vector of length \code{numNodes}, containing the index of vertex i
  in the ordering at location i.
}
\description{
Gives back the topological order of a DAG.
}
\details{
See
\url{https://lemon.cs.elte.hu/pub/doc/1.3.1/a00616.html#gafc2cb20cf3859f157c0e12da7f310bb3}
for more information.
}
