% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{IsAcyclic}
\alias{IsAcyclic}
\title{Check if Graph is Acyclic.}
\usage{
IsAcyclic(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's
edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's
edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
A \code{logical} stating if the graph is acyclic
}
\description{
A cycle is a path starting and ending in the same node and containing at
least one other node. A acyclic graph contains no cycles.
}
\details{
See \url{https://lemon.cs.elte.hu/pub/doc/1.3.1/a00616.html#ga14c191b2133a1dd23e1527f074c821c0}
for more information.
}
