% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{IsBiEdgeConnected}
\alias{IsBiEdgeConnected}
\title{Chcek if Graph is Bi-Edge-Connected}
\usage{
IsBiEdgeConnected(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's
edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's
edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
A \code{logical} stating if the graph is bi-edge connected
}
\description{
Checks if an undirected graph is bi-edge-connected, that is if there are no
edges that, if removed, would split the graph into two unconnected graphs.
}
\details{
See
\url{https://lemon.cs.elte.hu/pub/doc/1.3.1/a00616.html#ga37d22a2ddd5a064a9203720f2b93518e}
for more information.
}
