% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{IsBiNodeConnected}
\alias{IsBiNodeConnected}
\title{Checks if Graph is Bi-Node-Connected}
\usage{
IsBiNodeConnected(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's
edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's
edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
A \code{logical} stating if the graph is bi-node connected
}
\description{
Checks if an undirected graph is bi-node-connected, that is if there is are no
nodes which, if removed, would split the graph into two unconnected graphs.
}
\details{
See \url{https://lemon.cs.elte.hu/pub/doc/1.3.1/a00616.html#gac9257323ead7cbe64b7b4a628c4876b3}
for more information.
}
