% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{IsBipartite}
\alias{IsBipartite}
\title{IsBipartite}
\usage{
IsBipartite(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's
edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's
edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
A boolean stating if the graph is bipartite
}
\description{
Checks if an undirected graph is bipartite.
See 'https://lemon.cs.elte.hu/pub/doc/1.3.1/a00616.html#ga577db110d33bd487aaad5bfffb31c6f5'
for more information.
}
