% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{IsConnected}
\alias{IsConnected}
\title{Checks if a graph is connected}
\usage{
IsConnected(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{A vector of integers containing the source vertices of the edges}

\item{arcTargets}{A vector of integers containing the destination verctices of the edges}

\item{numNodes}{The number of nodes in the graph.}
}
\value{
A boolean stating if we have a connected graph or not
}
\description{
Checks if a graph is connected
}
