% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{IsConnected}
\alias{IsConnected}
\title{Check if Graph is Connected}
\usage{
IsConnected(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's
edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's
edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
A \code{logical} stating if the graph is connected
}
\description{
A connected graph has a path between any two nodes in the graph.
}
\details{
See \url{https://lemon.cs.elte.hu/pub/doc/1.3.1/a00616.html#gad5c8d1b650f6b614a852f8430d90e184}
for more information.
}
