% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{IsDAG}
\alias{IsDAG}
\title{Checks if a graph is a DAG}
\usage{
IsDAG(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{A vector of integers containing the source vertices of the edges}

\item{arcTargets}{A vector of integers containing the destination verctices of the edges}

\item{numNodes}{The number of nodes in the graph.}
}
\value{
A boolean stating if we have a DAG or not
}
\description{
Checks if a graph is a DAG
}
