% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{IsDAG}
\alias{IsDAG}
\title{Check if Graph is a DAG.}
\usage{
IsDAG(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's
edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's
edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
A \code{logical} stating if the graph is DAG
}
\description{
A graph is a DAG if it is Directed and Acyclic.
}
\details{
See \url{https://lemon.cs.elte.hu/pub/doc/1.3.1/a00616.html#gaef2b43c8cd1d74e15fa5c7607bc5e396}
for more information.
}
