% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{IsEulerian}
\alias{IsEulerian}
\title{IsEulerian}
\usage{
IsEulerian(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's
edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's
edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
A boolean stating if the graph is Eulerian
}
\description{
Checks if a graph is Eulerian. Works for both undirected and directed
graphs.
See 'https://lemon.cs.elte.hu/pub/doc/1.3.1/a00616.html#gafb5a4961cac4d877006869fc4cb6ea1d'
for more information.
}
