% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{IsLoopFree}
\alias{IsLoopFree}
\title{Checks if Graph is Loop Free}
\usage{
IsLoopFree(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's
edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's
edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
TRUE if the graph is loop free, FALSE otherwise
}
\description{
A loop is an edge that starts and ends at the same node and passes through
no other nodes.
}
\details{
See \url{https://lemon.cs.elte.hu/pub/doc/1.3.1/a00616.html#ga127f3963003cd532c79c226885fe1c8c}
for more information.
}
