% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{IsParallelFree}
\alias{IsParallelFree}
\title{IsParallelFree}
\usage{
IsParallelFree(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's
edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's
edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
A boolean stating if the graph is loop free
}
\description{
Checks if a graph has no parallel arcs/edges. Works for both undirected and
directed graphs.
See 'https://lemon.cs.elte.hu/pub/doc/1.3.1/a00616.html#gaa05e0683f90b69f31eb29fe7d09afde4'
for more information.
}
