% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{IsStronglyConnected}
\alias{IsStronglyConnected}
\title{IsStronglyConnected}
\usage{
IsStronglyConnected(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's
edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's
edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
A boolean stating if the graph is strongly connected
}
\description{
Checks if a directed graph is strongly connected.
See 'https://lemon.cs.elte.hu/pub/doc/1.3.1/a00616.html#gacd21b34d7b42b9835a204a57fcf15964'
for more information.
}
