% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{IsTree}
\alias{IsTree}
\title{IsTree}
\usage{
IsTree(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's edges}

\item{numNodes}{The
number of nodes in the graph}
}
\value{
A boolean stating if the graph is a tree
}
\description{
Checks if an undirected graph is a tree
}
