% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{IsTree}
\alias{IsTree}
\title{IsTree}
\usage{
IsTree(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's
edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's
edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
A boolean stating if the graph is a tree
}
\description{
Checks if an undirected graph is a tree.
See 'https://lemon.cs.elte.hu/pub/doc/1.3.1/a00616.html#gad1e4de234e926958647905478415bd54'
for more information.
}
