% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{IsTree}
\alias{IsTree}
\title{Check if Graph is a Tree}
\usage{
IsTree(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's
edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's
edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
A \code{logical} stating if the graph is a tree
}
\description{
A tree is an undirected graph in which any two nodes are connected by
exactly one path, or equivalently is both connected and acyclic.
}
\details{
See \url{https://lemon.cs.elte.hu/pub/doc/1.3.1/a00616.html#gad1e4de234e926958647905478415bd54}
for more information.
}
