% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxflow.R
\name{maxflow}
\alias{maxflow}
\title{MaxFlow}
\usage{
maxflow(
  arcSources,
  arcTargets,
  arcCapacities,
  sourceNode,
  destNode,
  numNodes,
  algorithm
)
}
\arguments{
\item{arcSources, }{a vector corresponding to the source nodes of a graph's edges}

\item{arcTargets, }{a vector corresponding to the destination nodes of a graph's edges}

\item{arcCapacities, }{a vector corresponding to the capacities of nodes of a graph's edges}

\item{sourceNode, }{the source node}

\item{destNode, }{the destination node}

\item{numNodes, }{the number of nodes in the graph}

\item{algorithm}{Which algorithm to run. Choices include
"Preflow", "EdmondsKarp",
where "Preflow" is the default.
See <http://lemon.cs.elte.hu/pub/doc/1.3.1/a00611.html> 
for details on the differences.}
}
\value{
A list containing three entries: 1) A list corresponding to the flows of arcs in the graph, and 2) A list of cut-values of the graph's nodes.
}
\description{
Runs maxflow algorithms on graphs.
}
