% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Minimum-Cost-Arborescence-Algorithms}
\alias{Minimum-Cost-Arborescence-Algorithms}
\alias{MinCostArborescenceRunner}
\title{Arborescence Algorithms}
\usage{
MinCostArborescenceRunner(
  arcSources,
  arcTargets,
  arcDistances,
  sourceNode,
  numNodes
)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's edges}

\item{arcDistances}{Vector corresponding to the distances of nodes of a graph's edges}

\item{sourceNode}{The source node}

\item{numNodes}{The number of nodes in the graph}
}
\value{
A list containing three entries: 1) Two vectors corresponding the source and target nodes of the edges in the tree, and 2) the total cost of the arborescence.
}
\description{
Arborescence Algorithms
}
