% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planarityembedding.R
\name{PlanarDrawing}
\alias{PlanarDrawing}
\title{Solver for Planar Drawing}
\usage{
PlanarDrawing(arcSources, arcTargets, numNodes)
}
\arguments{
\item{arcSources}{Vector corresponding to the source nodes of a graph's
edges}

\item{arcTargets}{Vector corresponding to the destination nodes of a graph's
edges}

\item{numNodes}{The number of nodes in the graph}
}
\value{
A named list of 1) "is_planar": a \code{logical} of if the graph is
  planar, 2) "x_coords": the x-coordinate of the planar embedding, 3)
  "y_coords": the y-coordinate of the planar embedding
}
\description{
The planar drawing algorithm calculates positions for the nodes in the
plane. These coordinates satisfy that if the edges are represented with
straight lines, then they will not intersect each other.
}
\details{
See \url{https://lemon.cs.elte.hu/pub/doc/1.3.1/a00307.html}
for more information.
}
